+++ noodles-sam/examples/sam_filter_data.rs
@@ -22,7 +22,7 @@ fn main() -> io::Result<()> {
     let mut args = env::args().skip(1);
     let src = args.next().expect("missing src");
 
-    let mut reader = sam::reader::Builder.build_from_path(src)?;
+    let mut reader = sam::reader::Builder::default().build_from_path(src)?;
     let header = reader.read_header()?;
 
     let stdout = io::stdout().lock();