+++ noodles-sam/src/io.rs
@@ -1,0 +1,5 @@
+//! SAM I/O.
+
+mod compression_method;
+
+pub use self::compression_method::CompressionMethod;